+++ lib/presentation/modules/onboarding/onboarding.dart
@@ -1,6 +1,8 @@
 library onborading;
 
+import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:groceries/config/localization/app_translations.dart';
@@ -14,5 +16,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 
 part 'views/onboarding_view.dart';
 part 'widgets/leave_onboarding_button.dart';
+part 'widgets/onboarding_form.dart';
 part 'widgets/onboarding_next_page_button.dart';
 part 'widgets/onboarding_page_base.dart';
+part 'cubit/onboarding_cubit.dart';
+part 'cubit/onboarding_state.dart';
